better TC module structure
