+++ tests/test_backend.py
@@ -1,18 +1,54 @@
 from datetime import timedelta
+from typing import Optional
 from unittest import TestCase
 from unittest.mock import MagicMock
 
+from spacepackets.ecss import PusTelecommand
 from tmtccmd import CcsdsTmtcBackend, CcsdsTmListener, TcHandlerBase
+from tmtccmd.com_if import ComInterface
 from tmtccmd.com_if.dummy import DummyComIF
 from tmtccmd.core import TcMode, TmMode, BackendRequest, BackendController
+from tmtccmd.tc import (
+    TcProcedureBase,
+    DefaultProcedureInfo,
+    TcProcedureType,
+    ProcedureCastWrapper,
+    TcQueueEntryBase,
+)
 from tmtccmd.tc.ccsds_seq_sender import SenderMode
+from tmtccmd.tc.handler import FeedWrapper
+
+
+class TcHandlerMock(TcHandlerBase):
+    def __init__(self):
+        super().__init__()
+        self.is_feed_cb_valid = False
+        self.feed_cb_call_count = 0
+
+    def send_cb(self, tc_queue_entry: TcQueueEntryBase, com_if: ComInterface):
+        pass
+
+    def queue_finished_cb(self, info: TcProcedureBase):
+        pass
+
+    def feed_cb(self, info: Optional[TcProcedureBase], wrapper: FeedWrapper):
+        self.feed_cb_call_count += 1
+        cast_wrapper = ProcedureCastWrapper(info)
+        if info is not None:
+            if info.ptype != TcProcedureType.DEFAULT:
+                def_info = cast_wrapper.to_def_procedure()
+                if def_info.service != "17":
+                    self.is_feed_cb_valid = False
+                elif def_info.op_code != "0":
+                    self.is_feed_cb_valid = False
+            wrapper.queue_helper.add_pus_tc(PusTelecommand(service=17, subservice=1))
 
 
 class TestCore(TestCase):
     def test_backend(self):
         com_if = DummyComIF()
         tm_listener = MagicMock(specs=CcsdsTmListener)
-        tc_handler = MagicMock(specs=TcHandlerBase)
+        tc_handler = TcHandlerMock()
         backend = CcsdsTmtcBackend(
             tc_mode=TcMode.IDLE,
             tm_mode=TmMode.IDLE,
@@ -55,3 +91,5 @@ class TestCore(TestCase):
         res = backend.periodic_op(ctrl)
         self.assertEqual(res.request, BackendRequest.TERMINATION_NO_ERROR)
         self.assertEqual(res.sender_res.mode, SenderMode.DONE)
+        backend.current_proc_info = DefaultProcedureInfo(service="17", op_code="0")
+        res = backend.periodic_op(ctrl)
